[feat] 추천 도서 컴포넌트 구현 및 분석 결과 컴포넌트 분리
